refactor: introduce zone providers and manager abstraction
