+++ src/dns_mw_zone.rs
@@ -1,42 +1,21 @@
-use std::borrow::Borrow;
-use std::collections::BTreeSet;
-use std::net::IpAddr;
-use std::str::FromStr;
-
-use crate::libdns::proto::op::Query;
-use crate::libdns::proto::rr::DNSClass;
-use crate::libdns::proto::rr::rdata::{PTR, TXT};
-
 use crate::config::HttpsRecordRule;
 use crate::dns::*;
-use crate::infra::{arp::lookup_client_mac_from_arp, ipset::IpSet};
 use crate::middleware::*;
+use crate::zone::{IdentityZoneProvider, LocalPtrZoneProvider, ZoneManager};
 
-const UNKNOWN_CLIENT_MAC: &str = "N/A";
+#[cfg(test)]
+use crate::libdns::proto::op::Query;
 
 pub struct DnsZoneMiddleware {
-    server_net: IpSet,
-    server_names: BTreeSet<Name>,
+    manager: ZoneManager,
 }
 
 impl DnsZoneMiddleware {
     pub fn new() -> Self {
-        let server_net = {
-            use local_ip_address::list_afinet_netifas;
-            let ips = list_afinet_netifas().unwrap_or_default();
-            IpSet::new(ips.into_iter().map(|(_, ip)| ip.into()))
-        };
-
-        let server_names = {
-            let mut set = BTreeSet::new();
-            set.insert(Name::from_str("smartdns.").unwrap());
-            set.insert(Name::from_str("whoami.").unwrap());
-            set
-        };
-
         Self {
-            server_net,
-            server_names,
+            manager: ZoneManager::new()
+                .with_provider(LocalPtrZoneProvider::new())
+                .with_provider(IdentityZoneProvider::new()),
         }
     }
 }
@@ -49,50 +28,11 @@ impl Middleware<DnsContext, DnsRequest, DnsResponse, DnsError> for DnsZoneMiddle
         req: &DnsRequest,
         next: Next<'_, DnsContext, DnsRequest, DnsResponse, DnsError>,
     ) -> Result<DnsResponse, DnsError> {
-        let query = req.query();
-        let name = query.name();
-        let query_type = query.query_type();
-
-        if let Some(response) = self.handle_builtin_txt_query(ctx, req) {
+        if let Some(response) = self.manager.lookup(ctx, req).await? {
             return Ok(response);
         }
 
-        match query_type {
-            RecordType::PTR => {
-                let mut is_current_server = false;
-                let name: &Name = name.borrow();
-
-                if self.server_names.contains(name) {
-                    is_current_server = true;
-                } else if let Ok(net) = name.parse_arpa_name() {
-                    is_current_server = self.server_net.overlap(&net);
-
-                    if !is_current_server {
-                        let is_private_ip = match net.addr() {
-                            IpAddr::V4(ip) => ip.is_private(),
-                            IpAddr::V6(ip) => {
-                                const fn is_unique_local(ip: std::net::Ipv6Addr) -> bool {
-                                    (ip.segments()[0] & 0xfe00) == 0xfc00
-                                }
-                                is_unique_local(ip)
-                            }
-                        };
-
-                        if is_private_ip {
-                            let mut res = DnsResponse::empty();
-                            res.add_query(query.original().to_owned());
-                            return Ok(res);
-                        }
-                    }
-                }
-
-                if is_current_server {
-                    return Ok(DnsResponse::from_rdata(
-                        req.query().original().to_owned(),
-                        RData::PTR(PTR(ctx.cfg().server_name())),
-                    ));
-                }
-            }
+        match req.query().query_type() {
             RecordType::SRV => {
                 if let Some(srv) = ctx.domain_rule.get_ref(|r| r.srv.as_ref()) {
                     return Ok(DnsResponse::from_rdata(
@@ -159,196 +99,6 @@ impl Middleware<DnsContext, DnsRequest, DnsResponse, DnsError> for DnsZoneMiddle
 
         next.run(ctx, req).await
     }
-}
-
-impl DnsZoneMiddleware {
-    fn handle_builtin_txt_query(&self, ctx: &DnsContext, req: &DnsRequest) -> Option<DnsResponse> {
-        let query = req.query().original().to_owned();
-
-        if query.query_type() != RecordType::TXT {
-            return None;
-        }
-
-        if !matches!(query.query_class(), DNSClass::CH | DNSClass::IN) {
-            return None;
-        }
-
-        let query_name = normalize_query_name(query.name());
-        let client_ip = normalize_client_ip(req.src().ip());
-        let server_name = trim_fqdn_dot(ctx.cfg().server_name().to_string());
-        let client_mac = || {
-            lookup_client_mac_from_arp(client_ip).unwrap_or_else(|| UNKNOWN_CLIENT_MAC.to_string())
-        };
-
-        match query_name.as_str() {
-            // BIND compatibility
-            "hostname.bind." | "id.server." => Some(txt_response(query, server_name.clone())),
-            "version.bind." => Some(txt_response(query, crate::BUILD_VERSION.to_string())),
-            "whoami.bind." | "client.ip.bind." | "clientip.bind." => {
-                Some(txt_response(query, client_ip.to_string()))
-            }
-            "whoami.mac.bind." | "client.mac.bind." | "clientmac.bind." => {
-                Some(txt_response(query, client_mac()))
-            }
-            "smartdns.info.bind." => Some(txt_response(
-                query,
-                build_info_kv_text(
-                    &server_name,
-                    crate::BUILD_VERSION,
-                    &client_ip,
-                    &client_mac(),
-                ),
-            )),
-            "smartdns.info.json.bind." => Some(txt_response(
-                query,
-                build_info_json_text(
-                    &server_name,
-                    crate::BUILD_VERSION,
-                    &client_ip,
-                    &client_mac(),
-                ),
-            )),
-            "smartdns.bind." => Some(txt_records_response(
-                query,
-                build_info_records_text(
-                    &server_name,
-                    crate::BUILD_VERSION,
-                    &client_ip,
-                    &client_mac(),
-                ),
-            )),
-
-            // SmartDNS native names (without `.bind`)
-            "hostname.smartdns." | "server-name.smartdns." | "hostname." => {
-                Some(txt_response(query, server_name.clone()))
-            }
-            "version.smartdns." | "server-version.smartdns." | "version." => {
-                Some(txt_response(query, crate::BUILD_VERSION.to_string()))
-            }
-            "whoami.smartdns." | "client-ip.smartdns." | "clientip.smartdns." | "whoami." => {
-                Some(txt_response(query, client_ip.to_string()))
-            }
-            "whoami-mac.smartdns." | "client-mac.smartdns." | "clientmac.smartdns." => {
-                Some(txt_response(query, client_mac()))
-            }
-            "info.smartdns." => Some(txt_response(
-                query,
-                build_info_kv_text(
-                    &server_name,
-                    crate::BUILD_VERSION,
-                    &client_ip,
-                    &client_mac(),
-                ),
-            )),
-            "json.smartdns." | "info.json.smartdns." => Some(txt_response(
-                query,
-                build_info_json_text(
-                    &server_name,
-                    crate::BUILD_VERSION,
-                    &client_ip,
-                    &client_mac(),
-                ),
-            )),
-            "smartdns." => Some(txt_records_response(
-                query,
-                build_info_records_text(
-                    &server_name,
-                    crate::BUILD_VERSION,
-                    &client_ip,
-                    &client_mac(),
-                ),
-            )),
-            _ => None,
-        }
-    }
-}
-
-fn normalize_query_name(name: &Name) -> String {
-    let mut normalized = name.clone();
-    normalized.set_fqdn(true);
-    normalized.to_string().to_ascii_lowercase()
-}
-
-fn trim_fqdn_dot(name: String) -> String {
-    name.trim_end_matches('.').to_string()
-}
-
-fn normalize_client_ip(ip: IpAddr) -> IpAddr {
-    match ip {
-        IpAddr::V6(addr) => addr.to_ipv4_mapped().map_or(IpAddr::V6(addr), IpAddr::V4),
-        IpAddr::V4(addr) => IpAddr::V4(addr),
-    }
-}
-
-fn txt_response(query: Query, value: String) -> DnsResponse {
-    let mut record = Record::from_rdata(
-        query.name().to_owned(),
-        crate::dns_client::MAX_TTL,
-        RData::TXT(TXT::new(vec![value])),
-    );
-    record.set_dns_class(query.query_class());
-    DnsResponse::new_with_max_ttl(query, vec![record])
-}
-
-fn txt_records_response(query: Query, values: Vec<String>) -> DnsResponse {
-    let records = values
-        .into_iter()
-        .map(|value| {
-            let mut record = Record::from_rdata(
-                query.name().to_owned(),
-                crate::dns_client::MAX_TTL,
-                RData::TXT(TXT::new(vec![value])),
-            );
-            record.set_dns_class(query.query_class());
-            record
-        })
-        .collect::<Vec<_>>();
-
-    DnsResponse::new_with_max_ttl(query, records)
-}
-
-fn build_info_kv_text(
-    server_name: &str,
-    version: &str,
-    client_ip: &IpAddr,
-    client_mac: &str,
-) -> String {
-    format!(
-        "server_name={server_name};server_version={version};client_ip={client_ip};client_mac={client_mac}",
-    )
-}
-
-fn build_info_records_text(
-    server_name: &str,
-    version: &str,
-    client_ip: &IpAddr,
-    client_mac: &str,
-) -> Vec<String> {
-    vec![
-        format!("server_name={server_name}"),
-        format!("server_version={version}"),
-        format!("client_ip={client_ip}"),
-        format!("client_mac={client_mac}"),
-    ]
-}
-
-fn build_info_json_text(
-    server_name: &str,
-    version: &str,
-    client_ip: &IpAddr,
-    client_mac: &str,
-) -> String {
-    let server_name = json_escape(server_name);
-    let version = json_escape(version);
-    let client_ip = json_escape(&client_ip.to_string());
-    let client_mac = json_escape(client_mac);
-    format!(
-        r#"{{"server_name":"{server_name}","server_version":"{version}","client_ip":"{client_ip}","client_mac":"{client_mac}"}}"#
-    )
-}
-
-fn json_escape(value: &str) -> String {
-    value.chars().flat_map(|c| c.escape_default()).collect()
 }
 
 #[cfg(test)]
@@ -494,7 +244,7 @@ mod tests {
             response.answers()[0]
                 .data()
                 .to_string()
-                .contains(UNKNOWN_CLIENT_MAC)
+                .contains("N/A")
         );
     }
