+++ src/zone/mod.rs
@@ -1,0 +1,7 @@
+mod manager;
+mod provider;
+mod providers;
+
+pub use manager::ZoneManager;
+pub use provider::ZoneProvider;
+pub use providers::{IdentityZoneProvider, LocalPtrZoneProvider};
