+++ src/zone/providers/mod.rs
@@ -1,0 +1,5 @@
+mod identity;
+mod local_ptr;
+
+pub use identity::IdentityZoneProvider;
+pub use local_ptr::LocalPtrZoneProvider;
